+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Restore topics spaces at startup
 
+- DuckDB implementation fixes.
+
 ## [1.2.2] - 2026-01-03
 
 ### Changed